コンポーネントの微調整
