+++ src/pages/letter_post/index.tsx
@@ -1,24 +1,13 @@
+import { NextPage } from "next";
 import { CommingSoonPages } from "@/components/elements/CommingSoonPages";
 import { LogoBikeistaButton } from "@/components/elements/LogoBikeistaButton";
-import { Box } from "@mui/material";
-import { NextPage } from "next";
+
 const LetterPost: NextPage = () => {
   return (
-    <>
-      <main>
-        <Box //ロゴボタンの基本位置
-          sx={{
-            position: "fixed", // 画面上部に固定位置に配置
-            marginTop: "40px",
-            marginLeft: "40px",
-            zIndex: 1000, // ボタンがテキストよりも手前に表示されるように
-          }}
-        >
-          <LogoBikeistaButton />
-        </Box>
-        <CommingSoonPages />
-      </main>
-    </>
+    <main>
+      <LogoBikeistaButton />
+      <CommingSoonPages />
+    </main>
   );
 };
