+++ src/pages/my_garage/index.tsx
@@ -1,9 +1,11 @@
+import { LogoBikeistaButton } from "@/components/elements/LogoBikeistaButton";
 import { NextPage } from "next";
 const MyGarage: NextPage = () => {
   return (
-    <>
-      <main>MyGarage</main>
-    </>
+    <main>
+      <LogoBikeistaButton />
+      MyGarage
+    </main>
   );
 };
